// definitions/omnicart-storefront/src/lib/content/scrapedProductContent.generated.ts
export const SCRAPED_PRODUCT_CONTENT = {
  "vnsh-quick-lock": {
    features: [
        {
            title: "Ultra Secure - Lightning Fast Access",
            description: "The simple but robust locking mechanism gives you total peace of mind that nobody is getting inside but you. But when seconds count, it gets your weapon in-hand lightning fast.",
            image: "/images/products/VNSH_Gun_Safe_4.jpg"
        },
        {
            title: "Great for Your Car, Nightstand, or Really Anywhere!",
            description: "This safe is the perfect size to tuck just about anywhere (or carry and take with you on the go). Small enough to easily fit in a bedroom drawer or cabinet, you can also toss it in your backpack then easily stow in any car glove box or console.",
            image: "/images/products/VNSH_Gun_Safe_3.jpg"
        },
        {
            title: "Not Just for Guns - Perfect for Other Valuables Too",
            description: "Not only is this sweet little safe great for securing your firearm… it’s also the ideal place to stash other valuables too. Use it to lock up cash, your personal IDs, watches and jewelry, or anything you don’t want swiped while you’re not looking!",
            image: "/images/products/VNSH_Gun_Safe_2.jpg"
        }
    ]
},
// ...
};